+++ crates/flagrant-api/src/handlers/variants.rs
@@ -12,7 +12,7 @@ use crate::{errors::ServiceError, extractors::DbConnection};
 /// - variant should be created for all environments (by default with same value)
 pub async fn create(
     DbConnection(mut conn): DbConnection,
-    Path((environment_id, feature_id)): Path<(u16, u16)>,
+    Path((environment_id, feature_id)): Path<(i32, i32)>,
     Json(payload): Json<VariantRequestPayload>,
 ) -> Result<Json<Variant>, ServiceError> {
     let env = environment::get_by_id(&mut conn, environment_id).await?;
@@ -26,7 +26,7 @@ pub async fn create(
 /// Updates existing variant with provided value/weight.
 pub async fn update(
     DbConnection(mut conn): DbConnection,
-    Path((environment_id, variant_id)): Path<(u16, u16)>,
+    Path((environment_id, variant_id)): Path<(i32, i32)>,
     Json(payload): Json<VariantRequestPayload>,
 ) -> Result<Json<()>, ServiceError> {
     let env = environment::get_by_id(&mut conn, environment_id).await?;
@@ -39,7 +39,7 @@ pub async fn update(
 
 pub async fn fetch(
     DbConnection(mut conn): DbConnection,
-    Path((environment_id, variant_id)): Path<(u16, u16)>,
+    Path((environment_id, variant_id)): Path<(i32, i32)>,
 ) -> Result<Json<Variant>, ServiceError> {
     let env = environment::get_by_id(&mut conn, environment_id).await?;
     let variant = variant::get_by_id(&mut conn, &env, variant_id).await?;
@@ -49,7 +49,7 @@ pub async fn fetch(
 
 pub async fn list(
     DbConnection(mut conn): DbConnection,
-    Path((environment_id, feature_id)): Path<(u16, u16)>,
+    Path((environment_id, feature_id)): Path<(i32, i32)>,
 ) -> Result<Json<Vec<Variant>>, ServiceError> {
     let env = environment::get_by_id(&mut conn, environment_id).await?;
     let feature = feature::get_by_id(&mut conn, &env, feature_id).await?;
@@ -60,7 +60,7 @@ pub async fn list(
 
 pub async fn delete(
     DbConnection(mut conn): DbConnection,
-    Path((environment_id, variant_id)): Path<(u16, u16)>,
+    Path((environment_id, variant_id)): Path<(i32, i32)>,
 ) -> Result<Json<()>, ServiceError> {
     let env = environment::get_by_id(&mut conn, environment_id).await?;
     let var = variant::get_by_id(&mut conn, &env, variant_id).await?;
